fix: pre-existing 2000-line hard-cap violations in service.py and cli_dump_helpers.py
Two files were over the AI-readiness 2000-line hard cap on main itself,
causing ai-readiness and unit-tests to fail on every CI run regardless
of what a PR touched:
- abicheck/service.py (2023 lines)
- abicheck/cli_dump_helpers.py (2001 lines)

Fixed with two file-size-cap-only splits, no behavior change:

- handle_non_elf_dump moved from cli_dump_helpers.py to a new sibling
  module, cli_dump_non_elf.py (mirroring the existing
  resolve_dump_depth -> cli_dump_depth.py precedent). cli.py now
  imports it from the new module directly.
- _attach_header_graph moved from service.py to a new sibling module,
  service_header_graph_attach.py (mirroring the existing
  service_scan.py/service_render.py precedent). Re-exported eagerly as
  service._attach_header_graph so the extensive
  monkeypatch.setattr("abicheck.service._attach_header_graph", ...)
  test surface and direct `from abicheck.service import
  _attach_header_graph` imports keep resolving unchanged.

scripts/check_ai_readiness.py's IMPORT_CYCLE_ALLOWLIST gained both new
modules as members of the existing, already-accepted CLI-registration/
service-routing strongly-connected component -- each is a split of an
already-member module, the identical shape service_compare_pipeline/
service_dump_pipeline/service_input_resolution were already signed off
under, not a new dependency direction. The one CLI_CONTRACT_ALLOWLIST
entry pinned to cli_dump_helpers.py's direct dumper.dump() call site
had its line number updated to match where that call now sits.

python scripts/check_ai_readiness.py now reports 0 errors (was 2).
Full isolated fast test suite: 30015 passed, 31 skipped, 4 xfailed,
0 unexpected failures.
