+++ scripts/check_ai_readiness.py
@@ -1282,6 +1282,20 @@ IMPORT_CYCLE_ALLOWLIST: frozenset[frozenset[str]] = frozenset(
                 "cli_debian_symbols",
                 "cli_doctor",
                 "cli_dump_helpers",
+                # `cli_dump_non_elf` joins the same SCC on exactly the terms
+                # `service_header_graph_attach` above was signed off under --
+                # a *split* of an existing member, not a new dependency
+                # direction. `handle_non_elf_dump` was previously defined
+                # directly in `cli_dump_helpers.py` (purely for the file-size
+                # cap, mirroring `resolve_dump_depth`/`resolve_dump_collect_
+                # context`'s earlier split into `cli_dump_depth.py`); it
+                # imports `_StampProvenance`/`_WriteSnapshotOutput` from
+                # `cli_dump_helpers` at module load (an edge internal to this
+                # already-allowlisted cluster, not a new one), and `cli`
+                # imports it back at module load in place of its former
+                # `cli_dump_helpers` import for the same name. No init
+                # deadlock -- the package still imports cleanly.
+                "cli_dump_non_elf",
                 "cli_graph",
                 "cli_helpers_compare",
                 "cli_inputs",
@@ -1311,6 +1325,23 @@ IMPORT_CYCLE_ALLOWLIST: frozenset[frozenset[str]] = frozenset(
                 "service",
                 "service_compare_pipeline",
                 "service_dump_pipeline",
+                # `service_header_graph_attach` joins the same SCC on exactly
+                # the terms `service_compare_pipeline`/`service_dump_pipeline`
+                # above were signed off under -- a *split* of an existing
+                # member, not a new dependency direction. `_attach_header_
+                # graph` was previously defined directly in `service.py`
+                # (purely for the file-size cap, mirroring the earlier splits
+                # of `service_render`/`service_scan` out of the same file); it
+                # imports `service_scan.expand_header_inputs` at module load
+                # (the identical edge `service.py` itself already carried via
+                # its own tail-of-file `from .service_scan import (...)`
+                # re-export block), and `service` imports it back eagerly at
+                # its own top. `service_scan` itself imports nothing from
+                # `service` at module load (it is a leaf) -- the SCC closes
+                # only through function-local imports elsewhere in this same
+                # cluster, so this adds no new *runtime* edge and no init
+                # deadlock; the package still imports cleanly.
+                "service_header_graph_attach",
                 "service_input_resolution",
                 "service_scan",
             }
@@ -2183,7 +2214,7 @@ CLI_CONTRACT_ALLOWLIST: frozenset[str] = frozenset(
         # Native ELF CLI dump (P0 item 2): calls `dumper.dump()` directly
         # rather than through `service_dump_pipeline.run_dump_request` —
         # tracked as Phase 1 item 1 of the convergence plan.
-        "abicheck/cli_dump_helpers.py:1679:19:dumper.dump",
+        "abicheck/cli_dump_helpers.py:1457:19:dumper.dump",
         # Standalone application-compatibility (P0 item 6): dumps both
         # sides directly rather than through any of the other paths.
         "abicheck/appcompat.py:1604:19:dumper.dump",
